refactor: split syntax extension helpers
